add: createAnswerUseCase

+++ src/modules/questions/useCases/DeleteQuestion/DeleteQuestionUseCase.ts
@@ -1,4 +1,5 @@
 import { IQuestionsRepository } from '@modules/questions/repositories/IQuestionsRepository';
+import { AppError } from '@shared/infra/http/errors/AppError';
 import { injectable, inject } from 'tsyringe';
 
 interface IRequest {
@@ -12,7 +13,15 @@ class DeleteQuestionUseCase {
     private questionsRepository: IQuestionsRepository
   ) {}
 
-  public async execute({ id }: IRequest): Promise<void> {
+  public async execute({ id }: IRequest): Promise<void | AppError> {
+
+    const questionExists = await this.questionsRepository.findById(id);
+
+    if(!questionExists) {
+        return new AppError("Question not exists!");
+    }
+
+
     await this.questionsRepository.delete(id);
 
     return;
